Changed pedagogicometer level to an enum instead of an integer, moved the Main method to Window1 to customise the start up of the splashscreen
Former-commit-id: 60690bf4f51bbd14f755d5e4a188fc0a2db66ab3

+++ MeTLMeeting/SandRibbon/Chrome/ToolBar.xaml.cs
@@ -35,7 +35,7 @@ namespace SandRibbon.Chrome
         }
         private void SetPedagogyLevel(PedagogyLevel level) 
         { 
-            foreach(var i in Enumerable.Range(0,level.code+1)){
+            foreach(var i in Enumerable.Range(0, ((int)level.code) + 1)){
                 switch (i)
                 {
                     case 0:
